pass for Projection for s-GTO

+++ utils.hpp
@@ -10,7 +10,7 @@
 					   std::complex<double> a,
 					   std::complex<double> b) {
 
-  typedef ::testing::internal::FloatingPoint<double> FP;
+  typedef ::testing::internal::FloatingPoint<float> FP;
   const FP a_r(a.real()), a_i(a.imag()), b_r(b.real()), b_i(b.imag());
 
   if(a_r.AlmostEquals(b_r) && a_i.AlmostEquals(b_i))
